refactor: 도메인로직에 해당하는 계산 수행 함수 계산기 모델 디렉토리로 이관
